// server/src/config/db-types.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;
// ...
export interface CalendarEvents {
  courseId: Generated<number | null>;
  createdAt: Generated<Date>;
  createdBy: number;
  description: Generated<string | null>;
  endsAt: Date;
  id: Generated<number>;
  isRecurring: Generated<boolean>;
  recurrenceEndDate: Generated<Date | null>;
  recurrenceFrequency: Generated<"daily" | "monthly" | "weekly" | null>;
  startsAt: Date;
  title: string;
  type: Generated<"class" | "exam" | "holiday" | "other">;
  updatedAt: Date;
}